AI interview practice partner/requirements.txt
streamlit
python-dotenv
google-genai
gTTS
fpdf
streamlit-mic-recorder